Build adjustments to mini_viz for MacOS X so it receives user input correctly.
On MacOS X programs can be GUI- or terminal-based but not both. If a
terminal-based program launches a GUI, then the terminal still gets the user
input (eg. key presses).

For mini_viz to work as expected on MacOS X then it must be labeled as
MACOSX_BUNDLE in CMake.